+++ src/background.ts
@@ -1,7 +1,8 @@
 import { NewExtensionServerMessage } from "./internal/utils/message";
-import { HttpUtils } from "./internal/utils/utils";
+import { HttpUtils, get } from "./internal/utils/utils";
 import { Application, Backend, Launcher } from "./internal/application";
 import { ConsoleLog } from "./internal/utils/log";
+import { SystemConfig } from "./internal/utils/config";
 
 
 class background implements Launcher {
@@ -17,6 +18,7 @@ class background implements Launcher {
             }
         });
         Application.CheckUpdate(function (isnew, data) {
+            let source = chrome.extension.getURL('src/mooc.js');
             if (isnew) {
                 chrome.browserAction.setBadgeText({
                     text: 'new'
@@ -25,6 +27,20 @@ class background implements Launcher {
                     color: [255, 0, 0, 255]
                 });
             }
+            //缓存js文件源码
+            let hotVersion = data.hotversion;
+            let littleVersion = hotVersion - data.version;
+            let isHotUpdate: boolean = false;
+            if (littleVersion < 0.01 && littleVersion > 0) {
+                Application.App.log.Info("使用热更新版本:" + hotVersion);
+                isHotUpdate = true;
+            }
+            if (isHotUpdate) {
+                source = SystemConfig.url + 'js/' + hotVersion + '.js';
+            }
+            get(source, function (source: string) {
+                chrome.storage.local.set({ "source": source });
+            });
         });
 
         let configKeyList: string[] = new Array();
@@ -34,7 +50,7 @@ class background implements Launcher {
         let configDefaultValue = new Map<string, any>().
             set("vtoken", "").set("rand_answer", false).set("auto", true).
             set("video_mute", true).set("answer_ignore", false).set("video_cdn", "").
-            set("video_multiple", 1).set("interval", 1);
+            set("video_multiple", 1).set("interval", 1).set("super_mode", false);
 
         chrome.storage.sync.get(configKeyList, function (items) {
             for (let key in items) {
