+++ src/mooc.ts
@@ -2,7 +2,7 @@ import { RemoveInjected } from "./internal/utils/utils";
 import { Application, Frontend, Launcher } from "./internal/application";
 import { ChromeConfigItems, NewFrontendGetConfig } from "./internal/utils/config";
 import { CreateMooc } from "./mooc/factory";
-import { ConsoleLog } from "./internal/utils/log";
+import { ConsoleLog, EmptyLog } from "./internal/utils/log";
 
 RemoveInjected(document);
