+++ src/views/popup.html
@@ -521,6 +521,10 @@
                 <input class="rdo" type="checkbox" id="answer-ignore" config-key="answer_ignore">
                 <label class="shou" for="answer-ignore" title="自动挂机时,忽略掉题目不做,直接跳过">忽略题目</label>
             </div>
+            <div class="config-item width_50">
+                <input class="rdo" type="checkbox" id="super-mode" config-key="super_mode">
+                <label class="shou" for="super-mode" title="超级模式,会有一些特殊的功能开启,安全性未知,具体请看使用文档">超级模式</label>
+            </div>
             <div>
                 <div class="config-item width_50">
                     <label for="video-cdn" title="播放源,为空默认选择最后一次播放源">播放源:</label>
