+++ webpack.dev.js
@@ -1,7 +1,7 @@
-const merge = require('webpack-merge');
-const common = require('./webpack.config.js');
+const merge = require("webpack-merge");
+const common = require("./webpack.config.js");
 
 module.exports = merge(common, {
     watch: true,
-    devtool: 'inline-source-map'
-})
+    devtool: "inline-source-map"
+});
