+++ webpack.prod.js
@@ -1,9 +1,6 @@
-const merge = require('webpack-merge');
-// const UglifyJSPlugin = require('uglifyjs-webpack-plugin');
-const common = require('./webpack.config.js');
+const merge = require("webpack-merge");
+const common = require("./webpack.config.js");
 
 module.exports = merge(common, {
-    plugins: [
-        // new UglifyJSPlugin()
-    ]
-})
+    plugins: []
+});
